+++ Kolonel-Backend/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "my-app",
+  "version": "0.1.0",
+  "private": true,
+  "main": "dist/index.js",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.1",
+    "express": "^4.18.1",
+    "mongoose": "^6.4.0",
+    "typescript": "^4.1.2"
+  },
+  "scripts": {
+    "build": "rimraf dist && tsc ",
+    "start": "node dist/index.js",
+    "dev": "tsc && nodemon dist/index.js"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/node": "^12.0.0",
+    "nodemon": "^2.0.16"
+  }
+}
