+++ Kolonel-frontend/public/favicon.ico
