+++ Kolonel-frontend/src/components/TodoForm.tsx
@@ -1,0 +1,28 @@
+import React, { useState, ChangeEvent, FormEvent } from 'react';
+
+interface TodoFormProps {
+  addTodo: AddTodo;
+}
+
+export const TodoForm: React.FC<TodoFormProps> = ({ addTodo }) => {
+  const [newTodo, setNewTodo] = useState<string>("");
+
+  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setNewTodo(e.target.value);
+  }
+
+  const handleSubmit = (e: FormEvent<HTMLButtonElement>) => {
+    e.preventDefault();
+    addTodo(newTodo);
+    setNewTodo("");
+  }
+
+  return (
+    <form className="todo-form">
+      <input type="text" value={newTodo} className="todo-input" placeholder="Add a todo" onChange={handleChange} />
+        <button type="submit" className="todo-button" onClick={handleSubmit}>
+          Add Todo
+        </button>
+    </form>
+  )
+};
