+++ Kolonel-frontend/src/index.css
@@ -1,0 +1,32 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+body {
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
+}
+
+.edit {
+    display: flex;
+    height: 100%;
+    width: 100%;
+}
+
+.edit form {
+    display: flex;
+    height: 100%;
+    width: 100%;
+}
+
+.edit form input {
+    display: flex;
+    height: 100%;
+    flex-grow: 1;
+    padding: 0.5rem;
+}
