--- ts modified/public/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta name="description" content="Acest website a fost creat folosind React" />
-    <!-- <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" /> -->
-    <!-- <link rel="manifest" href="%PUBLIC_URL%/manifest.json" /> -->
-    <title>Todo List</title>
-</head>
-
-<body>
-    <noscript>JavaScript nu a fost activat, pentru a vedea pagina web va rog să-l activați.</noscript>
-    <div id="root"></div>
-</body>
-
-</html>
