--- ts modified/src/App.css
@@ -1,147 +1,0 @@
-* {
-    font-family: 'Open Sans';
-}
-
-body {
-    background: linear-gradient( 110deg, rgb(20, 130, 204) 0%, rgb(251, 44, 44) 100%);
-}
-
-.todo-app {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    width: 500px;
-    min-height: 500px;
-    background: white;
-    text-align: center;
-    margin: 128px auto;
-    border-radius: 5px;
-}
-
-h1 {
-    margin: 32px 0;
-    font-size: 24px;
-}
-
-.completed {
-    text-decoration: line-through;
-    opacity: 0.4;
-}
-
-.todo-form {
-    margin-bottom: 32px;
-}
-
-.todo-button {
-    padding: 16px;
-    border: none;
-    border-radius: 0 20px 20px 0;
-    cursor: pointer;
-    outline: none;
-    background: linear-gradient(62deg, #667eea 0%, #764ba2 100%);
-    color: #fff;
-    text-transform: capitalize;
-}
-
-.todo-input {
-    padding: 15px 32px 15px 16px;
-    border-radius: 20px 0 0 20px;
-    border: 1px solid #dfe1e5;
-    outline: none;
-    width: 320px;
-    background: transparent;
-}
-
-.todo-input.edit {
-    border: 2px solid #149fff;
-}
-
-.todo-button.edit {
-    background: linear-gradient( 90deg, rgba(20, 159, 255, 1) 0%, rgba(17, 122, 255, 1) 100%);
-    padding: 16px 22px;
-}
-
-.edit-options {
-    color: white;
-    background: linear-gradient(62deg, #667eea 0%, #764ba2 100%);
-    padding: 5px 20px;
-    border: none;
-    border-radius: 5rem;
-    margin-left: 3px;
-    margin-right: 3px;
-}
-
-.todo-container {
-    display: flex;
-    flex-direction: row;
-    position: relative;
-}
-
-ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-}
-
-.todo-row {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
-    margin: 4px auto;
-    padding: 16px;
-    border-radius: 20px;
-    width: 80%;
-}
-
-.todo-row input {
-    margin-right: 10px;
-}
-
-.dropdown {
-    position: relative;
-    display: inline-block;
-    text-align: left;
-}
-
-.dropdown-list-container {
-    transform-origin: top;
-    position: absolute;
-    padding: 10px 0px;
-    right: 0rem;
-    margin-top: 2rem;
-    border-radius: 0.375rem;
-    box-shadow: 0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1);
-    background-color: white;
-}
-
-.dropdown-list-btn {
-    background-color: white;
-    border: none;
-    padding: 10px 30px;
-}
-
-.dropdown-list-btn:hover {
-    background-color: #eaeaea;
-}
-
-.edit-input {
-    border-radius: 3rem;
-    border: none;
-}
-
-.edit-input:focus {
-    border-radius: 3rem;
-    outline: none !important;
-    border-color: #764ba2;
-    box-shadow: 0 0 5px #764ba2;
-    /* box-shadow: inset 0 0 3px #764ba2, inset 0 0 3px #f50a1a; */
-}
-
-.red {
-    color: red;
-}
-
-.blue {
-    color: blue;
-}
